refactor: extract DataTab from ChartPropertiesPanel
Moves series/fit card management, the shared persistent config form,
add/remove, selection, and accordion expand/collapse into a new DataTab
widget, decoupled from StyleTab/AxesTab via seriesSelected/seriesListChanged/
axesRefreshRequested signals.

Also fixes a selection-tracking bug surfaced by verification: StyleTab's
"Chart" chip selection was getting clobbered on every Data-tab card rebuild
(including passive ones like theme refresh or accordion toggle) because the
chip-list refresh's "was Chart previously selected" check relied on a
ValueComboBox placeholder value that looked identical to a genuine prior
selection before any real chart had ever been loaded. StyleTab now tracks
this explicitly and takes the Data tab's selected index as an explicit
parameter instead of re-deriving it unreliably.
