+++ pandaplot/gui/components/sidebar/chart/chart_properties_panel.py
@@ -1,39 +1,25 @@
 """Chart properties side panel for configuring chart appearance and data."""
-from typing import List, Optional, override
+from typing import Optional, override
 
-from PySide6.QtCore import Qt
 from PySide6.QtWidgets import (
     QComboBox,
-    QFrame,
-    QGridLayout,
-    QHBoxLayout,
     QLabel,
-    QLineEdit,
-    QPushButton,
     QTabWidget,
     QVBoxLayout,
     QWidget,
 )
 
-from pandaplot.commands.project.chart import (
-    AddSeriesCommand,
-    ApplyChartPropertiesCommand,
-    RemoveFitDataCommand,
-    RemoveSeriesCommand,
-)
-from pandaplot.gui.components.common.card import Card
+from pandaplot.commands.project.chart import ApplyChartPropertiesCommand
 from pandaplot.gui.components.common.dirty_footer import DirtyFooter
-from pandaplot.gui.components.common.section_header import SectionHeader
-from pandaplot.gui.components.common.segmented_control import SegmentedControl
 from pandaplot.gui.components.sidebar.chart.tabs.axes_tab import AxesTab
 from pandaplot.gui.components.sidebar.chart.tabs.chart_tab import ChartTab
+from pandaplot.gui.components.sidebar.chart.tabs.data_tab import DataTab
 from pandaplot.gui.components.sidebar.chart.tabs.legend_tab import LegendTab
 from pandaplot.gui.components.sidebar.chart.tabs.style_tab import StyleTab
 from pandaplot.gui.core.widget_extension import PWidget
 from pandaplot.models.chart.chart_configuration import ChartType
 from pandaplot.models.events import ChartEvents, ProjectEvents, UIEvents
-from pandaplot.models.project.items import Dataset
-from pandaplot.models.project.items.chart import YAxis, restore_chart_state, snapshot_chart_state
+from pandaplot.models.project.items.chart import restore_chart_state, snapshot_chart_state
 from pandaplot.models.state.app_context import AppContext
 from pandaplot.services.theme.theme_manager import ThemeManager
 
@@ -55,29 +41,12 @@ class ChartPropertiesPanel(PWidget):
         self.command_executor = app_context.command_executor
         self.current_project = None
         self.current_chart = None  # Current Chart object being edited
-        self.datasets: List = []
         # Internal flags/state for safe UI updates
         self._updating_controls: bool = False  # Guard to prevent feedback loops
-        self._pending_label: str = ""        # Buffer while user types label
         self._has_unsaved_changes: bool = False
         # Baseline for Cancel and for Apply's undo: the chart state as of the
         # last load into this panel or the last Apply.
         self._loaded_snapshot: Optional[dict] = None
-        # Reference to the expanded Data-tab card's Y1/Y2 badge QLabel (and
-        # the design tokens it was last styled with), so a live series
-        # Y-axis edit can restyle it in place. See _on_series_config_changed.
-        self._expanded_card_y_axis_badge: Optional[QLabel] = None
-        self._expanded_card_y_axis_badge_tokens: dict = {}
-        # Which entry (data series index, then fit-data index appended after
-        # all series) is currently *selected* -- drives the Style tab's
-        # editing target and the live configuration form shown on the Data
-        # tab. Independent of `_expanded_card_indices` below: a card can be
-        # expanded (accordion open) without being selected.
-        self._expanded_series_index: int = 0
-        # Purely-visual accordion state: which cards show their expanded
-        # detail view. The selected card is always implicitly expanded (it
-        # hosts the live form) even if its index isn't in this set.
-        self._expanded_card_indices: set = {0}
 
         self._initialize()
         self._connect_signals()
@@ -113,7 +82,6 @@ class ChartPropertiesPanel(PWidget):
         # needs to connect directly to self.style_tab.set_chart_type here.
         self.style_tab = StyleTab(self.app_context, self)
         self.style_tab.configChanged.connect(self._on_any_tab_config_changed)
-        self.style_tab.seriesChipSelected.connect(self._expand_series)
 
         # Chart tab: chart identity (title, chart type, histogram bins)
         self.chart_tab = ChartTab(self)
@@ -129,11 +97,14 @@ class ChartPropertiesPanel(PWidget):
         self.axes_tab.configChanged.connect(self._on_any_tab_config_changed)
 
         # Data tab: series list + per-series dataset/X/Y/label configuration
-        data_tab = QWidget()
-        data_tab_layout = QVBoxLayout(data_tab)
-        self._create_series_management_section(data_tab_layout)
-        data_tab_layout.addStretch(1)
-        self.data_tab = data_tab
+        self.data_tab = DataTab(self.app_context, self)
+        self.data_tab.configChanged.connect(self._on_any_tab_config_changed)
+        self.data_tab.seriesSelected.connect(lambda kind, obj: self.style_tab.set_selected(kind, obj))
+        self.data_tab.seriesListChanged.connect(
+            lambda ds, fd: self.style_tab.set_series_list(ds, fd, self.data_tab.selected_index)
+        )
+        self.data_tab.axesRefreshRequested.connect(lambda: self.axes_tab.refresh_axis_chips(self.current_chart))
+        self.style_tab.seriesChipSelected.connect(self.data_tab._expand_series)
         self.tab_widget.addTab(self.data_tab, "Data")
 
         # Style tab (line/marker style)
@@ -253,9 +224,6 @@ class ChartPropertiesPanel(PWidget):
             }}
         """)
         
-        # Series management buttons
-        self._apply_series_button_styling()
-
         # Footer (DirtyFooter) theme token propagation
         tokens = theme_manager.get_design_tokens()
         self.footer.set_tokens(tokens)
@@ -264,12 +232,8 @@ class ChartPropertiesPanel(PWidget):
         # Style tab: shared widgets
         self.style_tab.apply_theme(tokens)
 
-        # Data tab: cards/SegmentedControl are rebuilt with fresh tokens
-        # every time _rebuild_series_cards runs, so re-running it here is
-        # the simplest way to make a live theme change reach them.
-        self._series_section_header.set_tokens(tokens)
-        self.series_y_axis_control.set_tokens(tokens)
-        self._rebuild_series_cards()
+        # Data tab: series management buttons/cards/SegmentedControl.
+        self.data_tab.apply_theme(tokens)
 
         # Axes tab: chip row plus each axis form's Card/SegmentedControl/
         # ToggleSwitch/SectionHeader widgets.
@@ -278,457 +242,19 @@ class ChartPropertiesPanel(PWidget):
         # Legend tab: shared widgets
         self.legend_tab.apply_theme(tokens)
 
-    def _apply_series_button_styling(self):
-        """Apply theme styling to series management buttons."""
-        theme_manager = self.app_context.get_manager(ThemeManager)
-        palette = theme_manager.get_surface_palette()
-        
-        # Get colors with fallbacks
-        accent = palette.get("accent", "#4CAF50")
-        secondary_fg = palette.get("secondary_fg", "#666666")
-        card_hover = palette.get("card_hover", "#e5f3ff")
-        base_fg = palette.get("base_fg", "#333333")
-
-        # Add series button (primary style)
-        add_style = f"""
-            QPushButton {{ 
-                background: {accent}; 
-                color: white; 
-                border: none; 
-                border-radius: 4px; 
-                padding: 4px 10px;
-            }}
-            QPushButton:hover {{ 
-                background: {card_hover}; 
-                color: {base_fg}; 
-            }}
-            QPushButton:disabled {{ 
-                background: {secondary_fg}; 
-            }}
-        """
-        self.add_series_button.setStyleSheet(add_style)
-        
-
-    def _create_series_management_section(self, layout):
-        """Create the data series management section: an expand/collapse card
-        per series (plus any fit-data entries), backed by a single persistent
-        configuration form that is reparented into whichever card is
-        currently expanded (tracked by `self._expanded_series_index`).
-        """
-        header_row = QHBoxLayout()
-        self._series_section_header = SectionHeader("Series")
-        header_row.addWidget(self._series_section_header)
-        header_row.addStretch(1)
-        self.add_series_button = QPushButton("+ Add series")
-        self.add_series_button.setCursor(Qt.CursorShape.PointingHandCursor)
-        self.add_series_button.clicked.connect(self._add_series)
-        header_row.addWidget(self.add_series_button)
-        layout.addLayout(header_row)
-
-        self._series_cards_container = QWidget()
-        self._series_cards_layout = QVBoxLayout(self._series_cards_container)
-        self._series_cards_layout.setContentsMargins(0, 0, 0, 0)
-        self._series_cards_layout.setSpacing(6)
-        layout.addWidget(self._series_cards_container)
-
-        # Persistent configuration form (dataset/X/Y/Y-axis/label). Created
-        # once so signal connections in _connect_signals stay valid for the
-        # panel's lifetime; it gets moved (reparented) into whichever card is
-        # currently expanded by _build_expanded_series_card.
-        self._series_form_widget = QWidget()
-        series_config_layout = QGridLayout(self._series_form_widget)
-        series_config_layout.setContentsMargins(0, 0, 0, 0)
-
-        series_config_layout.addWidget(QLabel("Dataset:"), 0, 0)
-        self.dataset_combo = QComboBox()
-        series_config_layout.addWidget(self.dataset_combo, 0, 1)
-
-        series_config_layout.addWidget(QLabel("X Column:"), 1, 0)
-        self.x_column_combo = QComboBox()
-        series_config_layout.addWidget(self.x_column_combo, 1, 1)
-
-        series_config_layout.addWidget(QLabel("Y Column:"), 2, 0)
-        self.y_column_combo = QComboBox()
-        series_config_layout.addWidget(self.y_column_combo, 2, 1)
-
-        series_config_layout.addWidget(QLabel("Y Axis:"), 3, 0)
-        self.series_y_axis_control = SegmentedControl(
-            [("Y₁ left", YAxis.PRIMARY), ("Y₂ right", YAxis.SECONDARY)]
-        )
-        series_config_layout.addWidget(self.series_y_axis_control, 3, 1)
-
-        series_config_layout.addWidget(QLabel("Label:"), 4, 0)
-        self.series_label_edit = QLineEdit()
-        series_config_layout.addWidget(self.series_label_edit, 4, 1)
-
-        self._rebuild_series_cards()
-
-    def _expand_series(self, index: int):
-        """Select `index` as the panel's live-edited entry: it drives the
-        Data tab's configuration form, the Style tab's target, and the
-        selected-card border highlight. Independent of any other card's
-        accordion open/closed state (`_expanded_card_indices`) -- see
-        `_toggle_card_expanded` for that purely-visual toggle."""
-        self._expanded_series_index = index
-        self._expanded_card_indices.add(index)
-        self._rebuild_series_cards()
-        if not self.current_chart:
-            return
-        if index < len(self.current_chart.data_series):
-            self.style_tab.set_selected("series", self.current_chart.data_series[index])
-        else:
-            self.style_tab.set_selected(
-                "fit", self.current_chart.fit_data[index - len(self.current_chart.data_series)]
-            )
-
-    def _toggle_card_expanded(self, index: int):
-        """Purely-visual accordion toggle: show/hide a card's read-only
-        detail view, independent of which card is *selected*."""
-        if index in self._expanded_card_indices:
-            self._expanded_card_indices.discard(index)
-        else:
-            self._expanded_card_indices.add(index)
-        self._rebuild_series_cards()
-
-    def _rebuild_series_cards(self):
-        """Rebuild the Data tab's card list from `self.current_chart`.
-
-        Each entry renders as one of three variants:
-        - the *selected* entry (`self._expanded_series_index`) always gets
-          the full configuration card (dataset/X/Y/Y-axis/label) -- it hosts
-          the one shared, live-wired form widget;
-        - other entries whose index is in `self._expanded_card_indices` get
-          a read-only detail row (purely visual "accordion open" state,
-          independent of selection);
-        - everything else gets the single-line collapsed chip row.
-
-        Safe to call at any point: fetches fresh theme tokens, so this
-        doubles as the mechanism by which cards pick up a live theme change
-        (see `_apply_theme`).
-        """
-        # Detach the persistent form widget from whatever card currently
-        # hosts it *before* that card is torn down below, so it survives
-        # the rebuild instead of being deleted as a child of a discarded card.
-        if self._series_form_widget.parent() is not None:
-            self._series_form_widget.setParent(None)
-
-        while self._series_cards_layout.count():
-            item = self._series_cards_layout.takeAt(0)
-            if item.widget():
-                item.widget().deleteLater()
-
-        theme_manager = self.app_context.get_manager(ThemeManager)
-        tokens = theme_manager.get_design_tokens()
-
-        if not self.current_chart:
-            if hasattr(self, "style_tab"):
-                self.style_tab.style_series_chips.clear()
-            self.axes_tab.refresh_axis_chips(self.current_chart)
-            return
-
-        total_series = len(self.current_chart.data_series)
-
-        for index, series in enumerate(self.current_chart.data_series):
-            if index == self._expanded_series_index:
-                card = self._build_expanded_series_card(index, tokens)
-            elif index in self._expanded_card_indices:
-                card = self._build_series_detail_row(series, index, tokens)
-            else:
-                card = self._build_collapsed_series_row(series, index, tokens)
-            self._series_cards_layout.addWidget(card)
-
-        for fit_offset, fit in enumerate(self.current_chart.fit_data):
-            index = total_series + fit_offset
-            if index == self._expanded_series_index:
-                card = self._build_expanded_series_card(index, tokens)
-            elif index in self._expanded_card_indices:
-                card = self._build_fit_detail_row(fit, index, tokens)
-            else:
-                card = self._build_collapsed_fit_row(fit, index, tokens)
-            self._series_cards_layout.addWidget(card)
-
-        self._refresh_style_chips()
-        self.axes_tab.refresh_axis_chips(self.current_chart)
-
-    def _refresh_style_chips(self):
-        """Sync the Style tab's dropdown with the same series+fit list the
-        Data tab's cards are built from, keeping its selection in lockstep
-        with `self._expanded_series_index` (the single, panel-wide
-        "currently edited entry" state also driven by Data-tab card expand)
-        -- unless "Chart" is the currently selected entry, which is its own
-        independent target and must survive a series/fit list refresh.
-
-        Values are the combined index (int) for series/fit, or the "chart"
-        sentinel, so selecting an entry can drive `_expand_series` directly
-        without needing to search for the clicked object's index.
-        """
-        if not hasattr(self, "style_tab"):
-            # The Data tab (built first) triggers an initial
-            # _rebuild_series_cards() call before the Style tab (and its
-            # dropdown) exists yet. There is no chart loaded yet at that
-            # point either, so there is nothing to reflect; the dropdown is
-            # populated for real the next time a chart is loaded.
-            return
-        style_chips = self.style_tab.style_series_chips
-        previous_value = style_chips.currentValue()
-        chip_items = [("Chart", "chart")]
-        for index, series in enumerate(self.current_chart.data_series):
-            label = series.label or f"{series.dataset_id}:{series.y_column}"
-            chip_items.append((label, index))
-        total_series = len(self.current_chart.data_series)
-        for fit_offset, fit in enumerate(self.current_chart.fit_data):
-            index = total_series + fit_offset
-            chip_items.append((f"\U0001f527 {fit.label}", index))
-
-        style_chips.blockSignals(True)
-        style_chips.clear()
-        for label, value in chip_items:
-            style_chips.addItem(label, value)
-        style_chips.blockSignals(False)
-
-        if previous_value == "chart":
-            style_chips.setCurrentValue("chart")
-        else:
-            style_chips.setCurrentValue(self._expanded_series_index)
-
-        # Keep the Style tab's own selection tracking (`_current_target`) in
-        # lockstep with the chip value set above, since `setCurrentValue`
-        # doesn't emit `currentValueChanged` (see ValueComboBox).
-        final_value = style_chips.currentValue()
-        if final_value == "chart":
-            self.style_tab._current_target = ("chart", None)
-            self.style_tab._update_target_cards_visibility()
-        elif final_value < len(self.current_chart.data_series):
-            self.style_tab.set_selected("series", self.current_chart.data_series[final_value])
-        else:
-            self.style_tab.set_selected(
-                "fit", self.current_chart.fit_data[final_value - len(self.current_chart.data_series)]
-            )
-
-    def _make_swatch(self, color: str, tokens: dict) -> QFrame:
-        swatch = QFrame()
-        swatch.setFixedSize(14, 14)
-        swatch.setStyleSheet(
-            f"background-color: {color}; "
-            f"border: 1px solid {tokens.get('border_control', '#999')}; "
-            f"border-radius: {tokens.get('radius_swatch', 4)}px;"
-        )
-        return swatch
-
-    def _build_trash_button(self, index: int, tokens: dict) -> QPushButton:
-        """Per-row delete icon (replaces the old single bottom Remove button
-        so a specific series/fit can be removed regardless of which entry
-        is selected/expanded)."""
-        button = QPushButton("\U0001f5d1")  # wastebasket emoji
-        button.setFlat(True)
-        button.setFixedWidth(24)
-        button.setCursor(Qt.CursorShape.PointingHandCursor)
-        button.setToolTip("Remove")
-        button.setStyleSheet(
-            "QPushButton { border: none; background: transparent; "
-            f"color: {tokens.get('text_muted', '#666')}; }} "
-            "QPushButton:hover { color: #dc3545; }"
-        )
-        button.clicked.connect(lambda _checked=False, i=index: self._remove_series_at(i))
-        return button
-
-    def _build_chevron_button(self, index: int, expanded: bool) -> QPushButton:
-        """Accordion toggle: purely visual expand/collapse, independent of
-        selection (see `_toggle_card_expanded`)."""
-        chevron = QPushButton("▾" if expanded else "▸")
-        chevron.setFlat(True)
-        chevron.setFixedWidth(24)
-        chevron.setCursor(Qt.CursorShape.PointingHandCursor)
-        chevron.clicked.connect(lambda _checked=False, i=index: self._toggle_card_expanded(i))
-        return chevron
-
-    def _install_select_on_click(self, card: QWidget, index: int):
-        """Clicking anywhere on a collapsed/detail card's background selects
-        it (moves the live-edited entry there), without affecting its own or
-        any other card's accordion open/closed state."""
-        def _handler(event, i=index):
-            if event.button() == Qt.MouseButton.LeftButton:
-                self._expand_series(i)
-            event.accept()
-        card.mousePressEvent = _handler
-        card.setCursor(Qt.CursorShape.PointingHandCursor)
-
-    def _build_collapsed_series_row(self, series, index: int, tokens: dict) -> QWidget:
-        """A chip-like collapsed row: color square, name, Y-axis badge, trash, chevron."""
-        card = Card()
-        card.set_tokens(tokens)
-        self._install_select_on_click(card, index)
-        row = QHBoxLayout(card)
-
-        row.addWidget(self._make_swatch(series.color, tokens))
-
-        name_label = QLabel(series.label or f"{series.dataset_id}:{series.y_column}")
-        name_label.setStyleSheet(f"color: {tokens.get('text_primary', '#000')};")
-        row.addWidget(name_label, 1)
-
-        row.addWidget(self._build_y_axis_badge(series.y_axis, tokens))
-        row.addWidget(self._build_trash_button(index, tokens))
-        row.addWidget(self._build_chevron_button(index, expanded=False))
-
-        return card
-
-    def _build_collapsed_fit_row(self, fit, index: int, tokens: dict) -> QWidget:
-        """Collapsed row for a fit-data entry (no Y-axis picker for fits)."""
-        card = Card()
-        card.set_tokens(tokens)
-        self._install_select_on_click(card, index)
-        row = QHBoxLayout(card)
-
-        row.addWidget(self._make_swatch(fit.color, tokens))
-
-        name_label = QLabel(f"\U0001f527 {fit.label}")  # wrench emoji
-        name_label.setStyleSheet(f"color: {tokens.get('text_primary', '#000')};")
-        row.addWidget(name_label, 1)
-
-        row.addWidget(self._build_trash_button(index, tokens))
-        row.addWidget(self._build_chevron_button(index, expanded=False))
-
-        return card
-
-    def _build_series_detail_row(self, series, index: int, tokens: dict) -> QWidget:
-        """Read-only detail view for a series card that's accordion-expanded
-        but not the currently *selected* entry -- purely visual, since the
-        one shared live-editable form can only live on the selected card."""
-        card = Card()
-        card.set_tokens(tokens)
-        self._install_select_on_click(card, index)
-        outer = QVBoxLayout(card)
-
-        header = QHBoxLayout()
-        header.addWidget(self._make_swatch(series.color, tokens))
-        name_label = QLabel(series.label or f"{series.dataset_id}:{series.y_column}")
-        name_label.setStyleSheet(f"color: {tokens.get('text_primary', '#000')};")
-        header.addWidget(name_label, 1)
-        header.addWidget(self._build_y_axis_badge(series.y_axis, tokens))
-        header.addWidget(self._build_trash_button(index, tokens))
-        header.addWidget(self._build_chevron_button(index, expanded=True))
-        outer.addLayout(header)
-
-        detail = QLabel(f"Dataset: {series.dataset_id}   X: {series.x_column}   Y: {series.y_column}")
-        detail.setStyleSheet(f"color: {tokens.get('text_muted', '#666')}; font-size: 10px;")
-        outer.addWidget(detail)
-
-        return card
-
-    def _build_fit_detail_row(self, fit, index: int, tokens: dict) -> QWidget:
-        """Read-only detail view for a fit card that's accordion-expanded but
-        not the currently selected entry (see `_build_series_detail_row`)."""
-        card = Card()
-        card.set_tokens(tokens)
-        self._install_select_on_click(card, index)
-        outer = QVBoxLayout(card)
-
-        header = QHBoxLayout()
-        header.addWidget(self._make_swatch(fit.color, tokens))
-        name_label = QLabel(f"\U0001f527 {fit.label}")
-        name_label.setStyleSheet(f"color: {tokens.get('text_primary', '#000')};")
-        header.addWidget(name_label, 1)
-        header.addWidget(self._build_trash_button(index, tokens))
-        header.addWidget(self._build_chevron_button(index, expanded=True))
-        outer.addLayout(header)
-
-        detail = QLabel(f"Fit: {fit.fit_type}   X: {fit.source_x_column}   Y: {fit.source_y_column}")
-        detail.setStyleSheet(f"color: {tokens.get('text_muted', '#666')}; font-size: 10px;")
-        outer.addWidget(detail)
-
-        return card
-
-    def _build_y_axis_badge(self, y_axis, tokens: dict) -> QLabel:
-        """Small 'Y₁'/'Y₂' badge, accented for the secondary axis."""
-        badge = QLabel()
-        self._apply_y_axis_badge_style(badge, y_axis, tokens)
-        return badge
-
-    def _apply_y_axis_badge_style(self, badge: QLabel, y_axis, tokens: dict):
-        """Set a Y-axis badge's text/style in place (shared by initial build
-        and live in-place refresh from `_on_series_config_changed`)."""
-        is_secondary = y_axis == YAxis.SECONDARY
-        badge.setText("Y₂" if is_secondary else "Y₁")
-        bg = tokens.get("y2_accent_bg") if is_secondary else tokens.get("surface_inset", "#eee")
-        fg = tokens.get("y2_accent") if is_secondary else tokens.get("text_muted", "#666")
-        badge.setStyleSheet(
-            f"background-color: {bg}; color: {fg}; "
-            f"border-radius: {tokens.get('radius_chip', 12)}px; "
-            "padding: 1px 8px; font-size: 10px; font-weight: 600;"
-        )
-
-    def _build_expanded_series_card(self, index: int, tokens: dict) -> QWidget:
-        """The expanded card for the currently *selected* entry: title + the
-        persistent config form, loaded with `index`'s values (a data-series
-        index, or a fit-data index appended after all series, matching the
-        combined indexing used throughout this panel). Rendered with an
-        accent border (via the "selected" dynamic property) to distinguish
-        it from unselected cards."""
-        card = Card()
-        card.set_tokens(tokens)
-        card.setProperty("selected", True)
-        card.style().unpolish(card)
-        card.style().polish(card)
-        outer = QVBoxLayout(card)
-
-        total_series = len(self.current_chart.data_series)
-        is_fit = index >= total_series
-        if is_fit:
-            fit = self.current_chart.fit_data[index - total_series]
-            title_text = f"\U0001f527 {fit.label}"
-        else:
-            series = self.current_chart.data_series[index]
-            title_text = series.label or f"{series.dataset_id}:{series.y_column}"
-
-        header = QHBoxLayout()
-        title_label = QLabel(title_text)
-        title_label.setStyleSheet(f"font-weight: 600; color: {tokens.get('text_primary', '#000')};")
-        header.addWidget(title_label, 1)
-        # Keep a reference so _on_series_config_changed can refresh this
-        # badge in place on a live Y-axis edit, without a full card rebuild
-        # (see that method's docstring for why a rebuild is unsafe there).
-        self._expanded_card_y_axis_badge = None
-        if not is_fit:
-            badge = self._build_y_axis_badge(series.y_axis, tokens)
-            self._expanded_card_y_axis_badge = badge
-            self._expanded_card_y_axis_badge_tokens = tokens
-            header.addWidget(badge)
-        header.addWidget(self._build_trash_button(index, tokens))
-        chevron = QPushButton("▾")  # ▾, indicates "currently expanded"
-        chevron.setFlat(True)
-        chevron.setFixedWidth(24)
-        chevron.setEnabled(False)
-        header.addWidget(chevron)
-        outer.addLayout(header)
-
-        outer.addWidget(self._series_form_widget)
-
-        if is_fit:
-            self._load_fit_into_controls(fit)
-        else:
-            self._reset_controls_for_series()
-            self._load_series_into_controls(series)
-
-        return card
 
     def _connect_signals(self):
-        """Connect widget signals."""
-        self.dataset_combo.currentTextChanged.connect(self._on_dataset_changed)
+        """Connect widget signals.
 
-        # Chart tab's own fields (title/subtitle/chart type/hist bins) are
-        # wired internally by ChartTab; Style tab's chart_style_card and
-        # Line/Marker card fields are wired internally by StyleTab.
-        # Axes tab: each axis form's widgets are wired directly to shared
-        # handlers in _build_axis_form_widgets, not here (the forms are
-        # built dynamically per-prefix, so there are no static
-        # self.x_*/self.y_*/self.y2_* attributes to hook up in one place).
-        # Connect series configuration change signals
-        self.x_column_combo.currentTextChanged.connect(self._on_series_config_changed)
-        self.y_column_combo.currentTextChanged.connect(self._on_series_config_changed)
-        self.series_y_axis_control.currentValueChanged.connect(self._on_series_config_changed)
-        # Defer label persistence to editingFinished to avoid disruptive refresh while typing
-        self.series_label_edit.textChanged.connect(self._on_label_typing)
-        self.series_label_edit.editingFinished.connect(self._on_label_committed)
+        Chart tab's own fields (title/subtitle/chart type/hist bins) are
+        wired internally by ChartTab; Style tab's chart_style_card and
+        Line/Marker card fields are wired internally by StyleTab; Data tab's
+        dataset/X/Y/Y-axis/label fields are wired internally by DataTab.
+        Axes tab: each axis form's widgets are wired directly to shared
+        handlers in _build_axis_form_widgets, not here (the forms are built
+        dynamically per-prefix, so there are no static self.x_*/self.y_*/
+        self.y2_* attributes to hook up in one place).
+        """
 
     def setup_event_subscriptions(self):
         """Set up event subscriptions for tab changes."""
@@ -743,7 +269,7 @@ class ChartPropertiesPanel(PWidget):
 
     def _ensure_datasets_loaded(self):
         """Populate datasets if empty (idempotent)."""
-        if not self.datasets and self.app_context.app_state.current_project:
+        if not self.data_tab.datasets and self.app_context.app_state.current_project:
             self.set_project(self.app_context.app_state.current_project)
 
     def _on_project_loaded(self, event_data):
@@ -802,136 +328,8 @@ class ChartPropertiesPanel(PWidget):
 
         update_type = event_data.get("update_type", "")
         if update_type in ["fit_added", "series_added", "series_removed"]:
-            self._rebuild_series_cards()
+            self.data_tab._rebuild_series_cards()
             self.logger.debug("Chart properties panel refreshed for update: %s", update_type)
-    
-    def _add_series(self):
-        """Add a new data series."""
-        if not self.current_chart:
-            return
-
-        # Create a new series with default values
-        dataset_id = self.dataset_combo.currentData() if self.dataset_combo.count() > 0 else ""
-        dataset_name = self.dataset_combo.currentText() if self.dataset_combo.count() > 0 else ""
-        x_column = self.x_column_combo.currentText() if self.x_column_combo.count() > 0 else ""
-        y_column = self.y_column_combo.currentText() if self.y_column_combo.count() > 0 else ""
-
-        if dataset_id and x_column and y_column:
-            command = AddSeriesCommand(
-                self.app_context,
-                chart_id=self.current_chart.id,
-                dataset_id=dataset_id,
-                x_column=x_column,
-                y_column=y_column,
-                label=f"{dataset_name}:{y_column}",
-                color=self._get_next_series_color(),
-            )
-            self.command_executor.execute_command(command)
-
-            # Select the newly added series
-            new_index = len(self.current_chart.data_series) - 1
-            self._expanded_series_index = new_index
-            self._expanded_card_indices.add(new_index)
-            self._rebuild_series_cards()
-
-    def _remove_series_at(self, index: int):
-        """Remove the data series or fit-data entry at the combined `index`
-        (data-series indices first, then fit-data indices appended after),
-        adjusting selection and accordion state for the index shift."""
-        if not self.current_chart:
-            return
-
-        total_series = len(self.current_chart.data_series)
-        total_items = total_series + len(self.current_chart.fit_data)
-        if index < 0 or index >= total_items:
-            return
-
-        if index < total_series:
-            command = RemoveSeriesCommand(
-                self.app_context,
-                chart_id=self.current_chart.id,
-                series_index=index,
-            )
-        else:
-            command = RemoveFitDataCommand(
-                self.app_context,
-                chart_id=self.current_chart.id,
-                fit_index=index - total_series,
-            )
-        self.command_executor.execute_command(command)
-
-        def _shift(i):
-            if i > index:
-                return i - 1
-            if i == index:
-                return None
-            return i
-
-        self._expanded_card_indices = {
-            shifted for shifted in (_shift(i) for i in self._expanded_card_indices)
-            if shifted is not None
-        }
-
-        remaining_items = len(self.current_chart.data_series) + len(self.current_chart.fit_data)
-        shifted_selected = _shift(self._expanded_series_index)
-        if shifted_selected is None:
-            shifted_selected = index
-        self._expanded_series_index = max(0, min(shifted_selected, max(remaining_items - 1, 0)))
-        self._expanded_card_indices.add(self._expanded_series_index)
-        self._rebuild_series_cards()
-
-    def _on_series_config_changed(self):
-        """Handle dataset / column configuration changes for the selected series.
-
-        Label changes are intentionally deferred to editingFinished handled by
-        _on_label_committed to avoid disruptive list refresh while typing.
-        """
-        if self._updating_controls or not self.current_chart:
-            return
-
-        current_row = self._expanded_series_index
-        if current_row < 0:
-            return
-
-        total_series = len(self.current_chart.data_series)
-        if current_row < total_series:
-            # Update data series (guard for safety)
-            if current_row >= len(self.current_chart.data_series):
-                return
-            series = self.current_chart.data_series[current_row]
-            if self.dataset_combo.currentData():
-                series.dataset_id = self.dataset_combo.currentData()
-            series.x_column = self.x_column_combo.currentText()
-            series.y_column = self.y_column_combo.currentText()
-            series.y_axis = self.series_y_axis_control.currentValue()
-
-            # Refresh the Axes-tab Y2 chip immediately so switching a series
-            # to the secondary axis is reflected without waiting for Apply
-            # or a full chart reload. This only touches the axis_chips
-            # SegmentedControl (not the Data-tab card list), so it's safe to
-            # call from here.
-            self.axes_tab.refresh_axis_chips(self.current_chart)
-
-            # Update the expanded card's own Y1/Y2 badge in place too.
-            # Deliberately NOT calling `_rebuild_series_cards()` from this
-            # handler: that tears down and rebuilds the Data-tab card list,
-            # including detaching/reattaching `_series_form_widget` (which
-            # hosts the very control that triggered this handler) - the same
-            # reentrancy hazard already fixed once for live-edit handlers
-            # touching the reparented series form widget. Updating the
-            # existing badge label in place avoids that entirely.
-            if getattr(self, "_expanded_card_y_axis_badge", None) is not None:
-                self._apply_y_axis_badge_style(
-                    self._expanded_card_y_axis_badge,
-                    series.y_axis,
-                    getattr(self, "_expanded_card_y_axis_badge_tokens", {}),
-                )
-        else:
-            # Fit data: columns/dataset not editable, ignore
-            return
-
-        self._has_unsaved_changes = True
-        self._update_status_indicator()
 
     def _on_any_tab_config_changed(self):
         if not self.current_chart:
@@ -950,105 +348,6 @@ class ChartPropertiesPanel(PWidget):
             change_count=1 if self._has_unsaved_changes else 0,
         )
 
-    def _load_series_into_controls(self, series):
-        """Load a data series into the configuration controls."""
-        # Enable all controls for series editing
-        previous_guard = self._updating_controls
-        self._updating_controls = True
-        try:
-            self._reset_controls_for_series()
-
-            # Set dataset
-            for i in range(self.dataset_combo.count()):
-                if self.dataset_combo.itemData(i) == series.dataset_id:
-                    self.dataset_combo.setCurrentIndex(i)
-                    break
-
-            # Repopulate column combos from the (possibly changed) dataset before
-            # selecting the series' columns, so stale entries from a previous
-            # dataset/column state (e.g. after undo/redo of a column rename)
-            # don't linger.
-            self._populate_column_combos(series.dataset_id)
-
-            # Set columns
-            x_index = self.x_column_combo.findText(series.x_column)
-            if x_index >= 0:
-                self.x_column_combo.setCurrentIndex(x_index)
-
-            y_index = self.y_column_combo.findText(series.y_column)
-            if y_index >= 0:
-                self.y_column_combo.setCurrentIndex(y_index)
-
-            # Set Y axis (primary/secondary). SegmentedControl.setCurrentValue
-            # doesn't emit currentValueChanged, so no signal-blocking needed.
-            self.series_y_axis_control.setCurrentValue(series.y_axis)
-
-            # Set label (block signals while populating)
-            self.series_label_edit.blockSignals(True)
-            self.series_label_edit.setText(series.label)
-            self.series_label_edit.blockSignals(False)
-            self._pending_label = series.label
-
-            self.style_tab.load_series_style(series)
-        finally:
-            self._updating_controls = previous_guard
-
-    def _load_fit_into_controls(self, fit):
-        """Load fit data into the configuration controls."""
-        previous_guard = self._updating_controls
-        self._updating_controls = True
-        try:
-            # For fit data, disable dataset/column controls since they're not editable
-            self.dataset_combo.setEnabled(False)
-            self.x_column_combo.setEnabled(False)
-            self.y_column_combo.setEnabled(False)
-            self.series_y_axis_control.setEnabled(False)
-
-            # Show fit info in the label (block signals)
-            self.series_label_edit.blockSignals(True)
-            self.series_label_edit.setText(fit.label)
-            self.series_label_edit.blockSignals(False)
-            self._pending_label = fit.label
-
-            self.style_tab.load_fit_style(fit)
-        finally:
-            self._updating_controls = previous_guard
-
-    def _on_label_typing(self, text: str):
-        """Buffer label text while user is typing without mutating the model."""
-        self._pending_label = text
-
-    def _on_label_committed(self):
-        """Persist buffered label to model after editing finishes.
-
-        Unlike the old QListWidget-backed list, the label text isn't shown
-        anywhere else while this entry is expanded (its card's title header
-        is only (re)built on the next _rebuild_series_cards call), so no
-        rebuild is needed here to avoid disruptive focus loss while typing.
-        """
-        if self._updating_controls or not self.current_chart:
-            return
-        current_row = self._expanded_series_index
-        if current_row < 0:
-            return
-        total_series = len(self.current_chart.data_series)
-        new_label = self._pending_label or self.series_label_edit.text()
-        if current_row < total_series:
-            if current_row < len(self.current_chart.data_series):
-                self.current_chart.data_series[current_row].label = new_label
-        else:
-            fit_index = current_row - total_series
-            if 0 <= fit_index < len(self.current_chart.fit_data):
-                self.current_chart.fit_data[fit_index].label = new_label
-        self._pending_label = new_label
-
-    def _reset_controls_for_series(self):
-        """Reset controls for editing regular data series."""
-        self.dataset_combo.setEnabled(True)
-        self.x_column_combo.setEnabled(True)
-        self.y_column_combo.setEnabled(True)
-        self.series_y_axis_control.setEnabled(True)
-
     def _clear_controls(self):
         """Reset panel controls to neutral defaults without touching any chart."""
         previous_guard = self._updating_controls
@@ -1058,82 +357,15 @@ class ChartPropertiesPanel(PWidget):
             self.chart_tab.clear()
             self.axes_tab.clear()
             self.legend_tab.clear()
-            self.series_label_edit.clear()
+            self.data_tab.clear()
         finally:
             self._updating_controls = previous_guard
-
-    def _get_next_series_color(self) -> str:
-        """Get the next color for a new series."""
-        colors = ["#1f77b4", "#ff7f0e", "#2ca02c", "#d62728", "#9467bd", 
-                 "#8c564b", "#e377c2", "#7f7f7f", "#bcbd22", "#17becf"]
-        
-        if not self.current_chart or not self.current_chart.data_series:
-            return colors[0]
-        
-        return colors[len(self.current_chart.data_series) % len(colors)]
 
     def set_project(self, project):
         """Set the current project."""
         self.current_project = project
-        self._update_datasets()
-    
-    def _update_datasets(self):
-        """Update the available datasets."""
-        self.dataset_combo.clear()
-        self.datasets = []
-        
-        if self.current_project:
-            # Iterate through all items in the project to find datasets
-            for item in self.current_project.get_all_items():
-                if isinstance(item, Dataset):
-                    self.dataset_combo.addItem(item.name, item.id)
-                    self.datasets.append(item)
-    
-    def _populate_column_combos(self, dataset_id):
-        """Fill the x/y column combos with the columns of the given dataset.
+        self.data_tab.set_project(project)
 
-        Signals are blocked while clearing/populating so this can safely be
-        called from contexts that don't want side effects from
-        currentTextChanged (e.g. while loading a series into controls).
-
-        Returns the list of columns populated (empty list if none).
-        """
-        if not dataset_id or not self.current_project:
-            return []
-        dataset = self.current_project.find_item(dataset_id)
-        if isinstance(dataset, Dataset) and dataset.data is not None:
-            columns = list(dataset.data.columns)
-            self.x_column_combo.blockSignals(True)
-            self.y_column_combo.blockSignals(True)
-            try:
-                self.x_column_combo.clear()
-                self.y_column_combo.clear()
-                for column in columns:
-                    self.x_column_combo.addItem(column)
-                    self.y_column_combo.addItem(column)
-            finally:
-                self.x_column_combo.blockSignals(False)
-                self.y_column_combo.blockSignals(False)
-            return columns
-        return []
-
-    def _on_dataset_changed(self):
-        """Handle dataset selection change."""
-        dataset_id = self.dataset_combo.currentData()
-        columns = self._populate_column_combos(dataset_id)
-
-        # Set defaults if possible
-        if columns:
-            self.x_column_combo.setCurrentIndex(0)
-            self.y_column_combo.setCurrentIndex(1 if len(columns) >= 2 else 0)
-
-        # setCurrentIndex() only emits currentTextChanged when the index
-        # actually changes, which it won't for indices auto-selected by
-        # _populate_column_combos while signals were blocked. Sync
-        # explicitly so the selected series' x_column/y_column never go
-        # stale relative to the combos.
-        self._on_series_config_changed()
-    
     def _on_apply(self):
         """Handle apply button click."""
         if not self.current_chart:
@@ -1185,12 +417,10 @@ class ChartPropertiesPanel(PWidget):
                 self.style_tab.load_chart_style(chart)
 
                 # Expand the first series/fit entry and rebuild the card list
-                # (this also (re)loads it into the config form controls).
-                self._expanded_series_index = 0
-                self._expanded_card_indices = {0}
-                self._rebuild_series_cards()
-                # (_rebuild_series_cards already loads series/fit 0's style
-                # into the Style tab controls via _build_expanded_series_card.)
+                # (this also (re)loads it into the config form controls, and
+                # loads series/fit 0's style into the Style tab via the
+                # seriesSelected signal chain).
+                self.data_tab.load(chart)
 
                 # Load configuration
                 self.chart_tab.load(chart)
@@ -1202,13 +432,10 @@ class ChartPropertiesPanel(PWidget):
         else:
             # Clear/default values
             self._clear_controls()
-            self._expanded_series_index = 0
-            self._expanded_card_indices = {0}
-            self._rebuild_series_cards()
-    
+
     def apply_to_chart(self, chart):
         """Apply current panel settings to a Chart object.
-        
+
         Args:
             chart: Chart object to update
         """
@@ -1221,16 +448,26 @@ class ChartPropertiesPanel(PWidget):
         self.axes_tab.apply_to(chart)
         self.legend_tab.apply_to(chart)
 
-        # Apply style updates to the currently selected series or fit data
-        current_row = self._expanded_series_index
+        # Data-tab-owned fields: re-asserts the selected series' y_axis, and
+        # creates a default series (dataset/x/y/label/y_axis) if none exist
+        # yet. Deliberately runs *before* the style application below (unlike
+        # the pre-Task-5 order) so that if this creates a new series, it
+        # still picks up the Style tab's currently-shown style via
+        # `apply_series_style_to` next -- previously that combined creation
+        # call baked color/line_width/marker_size in directly; routing it
+        # through the same style-apply path as any other series is more
+        # complete (also covers line_style/alpha/marker_shape/etc.) and
+        # avoids DataTab needing a reference to StyleTab.
+        self.data_tab.apply_to(chart)
+
+        # Apply style updates to the currently selected series or fit data.
+        current_row = self.data_tab.selected_index
         if current_row >= 0:
             total_series = len(chart.data_series)
 
             if current_row < total_series:
                 # Update data series
                 series = chart.data_series[current_row]
-
-                series.y_axis = self.series_y_axis_control.currentValue()
                 self.style_tab.apply_series_style_to(series)
 
                 self.logger.debug(
@@ -1248,25 +485,6 @@ class ChartPropertiesPanel(PWidget):
                         "Applied style to fit data %d: %s (color=%s)",
                         fit_index, fit.label, fit.color
                     )
-
-        # If no series exist but we have configuration, create a default series
-        if not chart.data_series:
-            dataset_id = self.dataset_combo.currentData()
-            dataset_name = self.dataset_combo.currentText()
-            x_column = self.x_column_combo.currentText()
-            y_column = self.y_column_combo.currentText()
-
-            if dataset_id and x_column and y_column:
-                chart.add_data_series(
-                    dataset_id=dataset_id,
-                    x_column=x_column,
-                    y_column=y_column,
-                    color=self.style_tab.line_color_row.currentColor(),
-                    line_width=self.style_tab.line_width_slider.value(),
-                    marker_size=self.style_tab.marker_size_slider.value(),
-                    label=f"{dataset_name}:{y_column}",
-                    y_axis=self.series_y_axis_control.currentValue()
-                )
 
         chart.update_modified_time()
